@@ -1,17 +1,11 @@
 import { defineConfig } from 'vite';
-import laravel, { refreshPaths } from 'laravel-vite-plugin';
+import laravel from 'laravel-vite-plugin';
 
 export default defineConfig({
     plugins: [
         laravel({
-            input: [
-                'resources/css/app.css',
-                'resources/js/app.js',
-            ],
-            refresh: [
-                ...refreshPaths,
-                'app/Livewire/**',
-            ],
+            input: ['resources/css/app.css', 'resources/js/app.js'],
+            refresh: true,
         }),
     ],
 });
